requirements.txt
Django>=4.2,<4.3
wagtail>=5.2,<5.3
python-decouple==3.8